+++ src/pymetropolis/metro_pipeline/__init__.py
@@ -1,4 +1,4 @@
 from .config import Config as Config
 from .file import MetroFile as MetroFile
-from .pipeline import run_pipeline as run_pipeline
+from .pipeline import MetroPipeline as MetroPipeline
 from .steps import Step as Step